pyproject.toml
[project]
name = "youtube-fishing-analyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.14.0",
    "google-api-python-client>=2.192.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.3",
]
